move CommunicationParams
