+++ hoymiles_modbus/datatypes.py
@@ -121,3 +121,21 @@ class PlantData:
     """Alarm indicator. True means that at least one inverter reported an alarm."""
     inverters: list[InverterData] = field(default_factory=list)
     """Data for each inverter."""
+
+
+@dataclass
+class CommunicationParams:
+    """Low level pymodbus communication parameters."""
+
+    timeout: float = 3
+    """Timeout for a connection request, in seconds."""
+    retries: int = 3
+    """Max number of retries per request."""
+    reconnect_delay: float = 0
+    """Minimum delay in seconds.milliseconds before reconnecting.
+    Doubles automatically with each unsuccessful connect, from
+    **reconnect_delay** to **reconnect_delay_max**.
+
+    Default is 0 which means that reconnecting is disabled."""
+    reconnect_delay_max: float = 300
+    """Maximum delay in seconds.milliseconds before reconnecting."""
